ImgMath: move to package net.imglib2.algorithm.math
